+++ crates/biome_cli/src/commands/rage.rs
@@ -263,6 +263,7 @@ impl Display for RageConfiguration<'_, '_> {
                             {KeyValuePair("Indent size", markup!({DebugDisplayOption(json_formatter_configuration.indent_size)}))}
                             {KeyValuePair("Line ending", markup!({DebugDisplayOption(json_formatter_configuration.line_ending)}))}
                             {KeyValuePair("Line width", markup!({DebugDisplayOption(json_formatter_configuration.line_width.map(|lw| lw.get()))}))}
+                            {KeyValuePair("Trailing Comma", markup!({DebugDisplayOption(json_formatter_configuration.trailing_comma)}))}
                         ).fmt(fmt)?;
                     }