+++ crates/biome_json_formatter/src/json/lists/member_list.rs
@@ -9,13 +9,12 @@ pub(crate) struct FormatJsonMemberList;
 impl FormatRule<JsonMemberList> for FormatJsonMemberList {
     type Context = JsonFormatContext;
     fn fmt(&self, node: &JsonMemberList, f: &mut JsonFormatter) -> FormatResult<()> {
-        let options = f.options();
-        let allow_trailing_separator = options.allows_trailing_separator();
+        let trailing_separator = f.options().to_trailing_separator();
         let mut join = f.join_nodes_with_soft_line();
 
         for (element, formatted) in node
             .elements()
-            .zip(node.format_separated(",", allow_trailing_separator))
+            .zip(node.format_separated(",", trailing_separator))
         {
             join.entry(element.node()?.syntax(), &formatted);
         }